For terrain

+++ Project/Mesh.h
@@ -134,7 +134,7 @@ public:
 class CHeightMapImage
 {
 private:
-	BYTE							*m_pHeightMapPixels;
+	unsigned short					*m_pHeightMapPixels;
 
 	int								m_nWidth;
 	int								m_nLength;
@@ -148,7 +148,7 @@ public:
 	XMFLOAT3 GetHeightMapNormal(int x, int z);
 	XMFLOAT3 GetScale() { return(m_xmf3Scale); }
 
-	BYTE *GetHeightMapPixels() { return(m_pHeightMapPixels); }
+	unsigned short *GetHeightMapPixels() { return(m_pHeightMapPixels); }
 	int GetHeightMapWidth() { return(m_nWidth); }
 	int GetHeightMapLength() { return(m_nLength); }
 };
